Use `wp-error` for common Erorr handling logic
package: add "builtin-status-codes" as a dependency

make `toTitle()` work with spaces

So that "Not Found" will get condensed into "NotFound"

set a descriptive `name` and `message` for HTTP status errors

ensure that "Error" is only present once in the `name`

This is specifically needed for 500 "Internal Server Error".
We want "InternalServerError", not "InternalServerErrorError".

use "uppercamelcase" module for generating the error name

refactor to use "wp-error" module

update for `wp-error@1.1.0`

Simplifies the logic quite a bit